Move the StaticLocationEngine; didn't realize that was only in the example namespace
